requirements.txt
openai==1.3.0
anthropic==0.25.0
tavily-python==0.3.0
elevenlabs==0.2.28
python-dotenv==1.0.0
transformers==4.35.0
datasets==2.14.6
huggingface_hub==0.19.4
fastapi==0.104.1
uvicorn==0.24.0
pydub==0.25.1
requests==2.31.0
torch==2.2.0